Use Redux Toolkit
@reduxjs/toolkit 적용
